Makes the Trainer weights longs instead of ints.
This fixes an integer overflow in the Trainer: weights used to be word
scores (low numbers), but have been switched to differences between
timestamps in seconds (potentially very large).

Also adds a new Random class that extends java.util.Random, adding a
nextLong(n) method.

+++ src/ro/undef/patois/Database.java
@@ -350,7 +350,7 @@ public class Database {
 
         try {
             while (cursor.moveToNext())
-                weights.add(new Trainer.Weight(cursor.getLong(0), cursor.getInt(1)));
+                weights.add(new Trainer.Weight(cursor.getLong(0), cursor.getLong(1)));
         } finally {
             cursor.close();
         }
